% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fieller.R
\name{Fieller}
\alias{Fieller}
\title{A function for obtaining the confidence interval of a ratio by the Fieller's method}
\usage{
Fieller(mu_n,mu_d,var_n,var_d,rho=0,df_n=NULL,df_d=NULL,con_level=0.95,
              truncation_interval=FALSE,truncation_lower=NULL,
              truncation_upper=NULL,description=FALSE,H0_test=FALSE,DC_0)
}
\arguments{
\item{mu_n}{The estimated mean of the numerator.}

\item{mu_d}{The estimated mean of the denominator.}

\item{var_n}{A positive value gives the estimated variance of the numerator.}

\item{var_d}{A positive value gives the estimated variance of the denominator.}

\item{rho}{A value between -1 and 1 that represents the estimated correlation coefficient of the numerator and the denominator.}

\item{df_n}{The degree of freedom of the numerator. The default value is NULL.}

\item{df_d}{The degree of freedom of the denominator. The default value is NULL.}

\item{con_level}{The confidence level. Should be between 0 and 1. The default is 0.95.}

\item{truncation_interval}{According to the literature or known knowledge, the ratio should be within a certain interval, for example in the interval from 1 to 2, so we need to truncate the point estimate and the Fieller's confidence interval into the certain interval to get the final results. The default value is FALSE, that is, no truncation.}

\item{truncation_lower}{If truncation_interval=TRUE, you need to specify the truncation interval, and truncation_lower indicates the lower limit of the truncation interval. The default value is NULL.}

\item{truncation_upper}{If truncation_interval=TRUE, you need to specify the truncation interval, and truncation_upper indicates the upper limit of the truncation interval. The default value is NULL.}

\item{description}{Whether to describe the results of point_estimate and confidence interval in text. The default value is FASLE.}

\item{H0_test}{Whether to test the null hypothesis of no DC, incomplete DC or complete DC. The default value is FASLE.}

\item{DC_0}{If H0_test=TRUE, you need to specify the null hypothesis, where DC_0 takes the values of 1 and 2 for no DC or complete DC, respectively. The default value is NULL.}
}
\value{
a data.frame (H0_test=FALSE) or list (H0_test=TRUE):
\describe{
\item{point_estimate}{The point estimate of the ratio.}
\item{Lower_CL}{The lower bound of the estimated interval of the Fieller's method.}
\item{Upper_CL}{The upper bound of the estimated interval of the Fieller's method.}
\item{CI}{The estimated interval of the Fieller's method.}
\item{length_CI}{The length of the estimated confidence interval.}
\item{CI_type}{The CI of the Fieller's method can be divided into 'continuous interval','discontinuous interval', 'empty set', 'noninformative interval' or 'point'.}
\item{H0_test}{If H0_test=TRUE, returns the Z-score test statistic and corresponding p-value under the null hypothesis.}
}
}
\description{
This function is used to calculate the upper and lower limits of the Fieller's confidence interval for the ratio estimate.
}
\examples{
data(Graves_Meta_analysis)
attach(Graves_Meta_analysis)
SNP_AAf <- ifelse(rs3827440==2 & Sex==2,1,0)
SNP_Af <- ifelse(rs3827440!=0 & Sex==2,1,0)
SNP_m <- ifelse(rs3827440!=0 & Sex==1,1,0)
myfit <- glm(Graves_disease~Sex+SNP_AAf+SNP_Af+SNP_m,
            family = binomial(link = "logit"),
            na.action = na.exclude)
detach(Graves_Meta_analysis)
ncoef <- length(myfit$coefficients)
beta_hat <- coefficients(myfit)[(ncoef-2):ncoef]
cov_hat <- vcov(myfit)[(ncoef-2):ncoef,(ncoef-2):ncoef]

mu_n <- beta_hat[3]
mu_d <- (beta_hat[1]+beta_hat[2])/2
var_n <- cov_hat[3,3]
var_d <- (cov_hat[1,1]+cov_hat[2,2]+2*cov_hat[1,2])/4
cov_nd <- (cov_hat[1,3]+cov_hat[2,3])/2
rho <- cov_nd/sqrt(var_n*var_d)

Fieller(mu_n,mu_d,var_n,var_d,rho,df_n=NULL,df_d=NULL,con_level=0.95,
       truncation_interval=FALSE,truncation_lower=NULL,truncation_upper=NULL,
       description=FALSE,H0_test=TRUE,DC_0=c(1,2))

}
\references{
Wang, P.; Zhang, Y.; Wang, BQ; et al. A statistical measure for the skewness of X chromosome inactivation based on case-control design. BMC Bioinformatics 2019, 20, 11.
}
