% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequen_DC.R
\name{Frequen_DC}
\alias{Frequen_DC}
\title{A function for estimating the degree of dosage compensation at X-linked loci using the delta, Fieller's or penalized Fieller's method}
\usage{
Frequen_DC(SNP,Y,trait_type='quantitative',Sex,Covariate=NULL,
           variance_heterogeneity=FALSE,MGC_Cutoff=20,method='Fieller',
           df_n=NULL,df_d=NULL,con_level=0.95,truncation_interval=FALSE,
           truncation_lower=NULL,truncation_upper=NULL,description=FALSE,
           H0_test=FALSE,DC_0)
}
\arguments{
\item{SNP}{Each genotype is coded as 0, 1 or 2 for females and coded as 0 or 1 for males, indicating the number of reference allele. The length of \code{SNP} should be the same to those of \code{Y}, \code{Sex} and \code{Covariate}.}

\item{Y}{A numeric vector of a quantitative trait, such as human height.}

\item{trait_type}{A character string either being "quantitative" or "qualitative", the default value is "quantitative".}

\item{Sex}{A vector of the genetic sex following PLINK default coding, where males are coded as 1 and females are coded as 2.}

\item{Covariate}{Optional: a vector or a matrix of covariates, such as age and BMI.}

\item{variance_heterogeneity}{For quantitative traits, we empolyed weighted least squares to fit the model when variance_heterogeneity is set to TRUE, and ordinary least squares when variance_heterogeneity is set to FASLE.}

\item{MGC_Cutoff}{Cutoff for the minimum genotype count in either females or males (default=20). When the minimum genotype count for a SNP is below this specified cutoff, we will employ ordinary least squares instead of weighted least squares for genetic association analysis with quantitative traits. This is mainly to avoid inflated type I errors caused by the instability of the weights. (Yang et al., 2022; Deng et al., 2019; Soave et al., 2015).}

\item{method}{a character string indicating which kind of estimation methods is to be used. One of "Fieller"(default), "PenFieller" , "delta" or "all": can be abbreviated.}

\item{df_n}{The degree of freedom of the numerator. The default value is NULL.}

\item{df_d}{The degree of freedom of the denominator. The default value is NULL.}

\item{con_level}{The confidence level. Should be between 0 and 1. The default is 0.95.}

\item{truncation_interval}{According to the literature or known knowledge, the ratio should be within a certain interval, for example in the interval from 1 to 2, so we need to truncate the point estimate and the Fieller's confidence interval into the certain interval to get the final results. The default value is FALSE, that is, no truncation.}

\item{truncation_lower}{If truncation_interval=TRUE, you need to specify the truncation interval, and truncation_lower indicates the lower limit of the truncation interval. The default value is NULL.}

\item{truncation_upper}{If truncation_interval=TRUE, you need to specify the truncation interval, and truncation_upper indicates the upper limit of the truncation interval. The default value is NULL.}

\item{description}{Whether to describe the results of point_estimate and confidence interval in text. The default value is FASLE.}

\item{H0_test}{Whether to test the null hypothesis of no DC, incomplete DC or complete DC. The default value is FASLE.}

\item{DC_0}{If H0_test=TRUE, you need to specify the null hypothesis, where DC_0 takes the values of 1 and 2 for no DC or complete DC, respectively. The default value is NULL.}
}
\value{
a list:
\describe{
\item{point_estimate}{The point estimate of the ratio.}
\item{Lower_CL}{The lower bound of the estimated interval of the Fieller's method.}
\item{Upper_CL}{The upper bound of the estimated interval of the Fieller's method.}
\item{CI}{The estimated interval of the Fieller's method.}
\item{length_CI}{The length of the estimated confidence interval.}
\item{CI_type}{The CI of the Fieller's method can be divided into 'continuous interval','discontinuous interval', 'empty set', 'noninformative interval' or 'point'.}
\item{H0_test}{If H0_test=TRUE, returns the Z-score test statistic and corresponding p-value under the null hypothesis.}
}
}
\description{
This function is used to calculate the upper and lower limits of the delta, Fieller's or penalized Fieller's confidence interval for the ratio estimate.
}
\examples{
data("Graves_Meta_analysis")
Frequen_DC(Graves_Meta_analysis$rs3827440,Graves_Meta_analysis$Graves_disease,
           trait_type='qualitative',Graves_Meta_analysis$Sex,method='all',
           con_level=0.95,truncation_interval=TRUE,truncation_lower=1,truncation_upper=2,
           description=TRUE,H0_test=TRUE,DC_0=2)

}
\references{
Yang ZY, Liu W, Yuan YX, et al. Robust association tests for quantitative traits on the X chromosome. \emph{Heredity}, 2022, \strong{129}: 244–256.

Deng WQ, Mao S, Kalnapenkis A, et al. Analytical strategies to include the X-chromosome in variance heterogeneity analyses: evidence for trait-specific polygenic variance structure. \emph{Genetic Epidemiology}, 2019, \strong{43}: 815-830.

Soave D, Corvol H, Panjwani N, et al. A joint location-scale test improves power to detect associated SNPs, gene sets, and pathways. \emph{The American Journal of Human Genetics}, 2015, \strong{97}: 125–138.
}
